pyproject.toml
[project]
name = "snipster-app"
version = "0.1.6"
description = "A lightweight code snippet management system with CLI, API, and web interface"
readme = "README.md"
authors = [
    { name = "sl02", email = "" }
]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "sqlmodel>=0.0.27",
    "python-decouple==3.8",
    "typer>=0.21.0",
    "rich>=14.2.0",
    "fastapi[all]>=0.128.0",
    "streamlit>=1.53.0",
]

[project.scripts]
snipster = "snipster.cli:app"
snipster-web = "snipster.streamlit_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.coverage.run]
omit = [
    "src/snipster/api/main.py",
    "src/snipster/streamlit_app.py",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.workspace]
members = ["snipster-sl"]

[tool.hatch.build.targets.wheel]
packages = ["src/snipster"]
